Note that 'onFinalize' is run after the observer, not the input.
This is not necessarily a bug, but does result in a deadlock in
unconventional observer pipes. If the observer runs
independently of the input, then it could pull an element and close
the channel before the guard is released.

It might be worth modelling two different types of observers -
those that are directly linked to the input and those that aren't.